requirements.txt
fasttext==0.9.2
spacy==3.7.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
lightgbm==4.5.0
python-docx==1.1.2
PyPDF2==3.0.1
rapidfuzz==3.9.7
metaphone==0.6
matplotlib==3.9.2
openpyxl==3.1.5
